Synchronize workspace edits across windows
